+++ backend/CourtVision/src/main/java/cit/edu/capstone/CourtVision/dto/PhysicalUpdateRequestDTO.java
@@ -11,10 +11,10 @@ public class PhysicalUpdateRequestDTO {
     private String coachName;
     private Long teamId;
     private String teamName;
-    private BigDecimal weight;
-    private BigDecimal height;
-    private BigDecimal wingspan;
-    private BigDecimal vertical;
+    private double weight;
+    private double height;
+    private double wingspan;
+    private double vertical;
     private LocalDate dateRequested;
     private int requestStatus;
 
@@ -39,17 +39,17 @@ public class PhysicalUpdateRequestDTO {
     public String getTeamName() { return teamName; }
     public void setTeamName(String teamName) { this.teamName = teamName; }
 
-    public BigDecimal getWeight() { return weight; }
-    public void setWeight(BigDecimal weight) { this.weight = weight; }
+    public double getWeight() { return weight; }
+    public void setWeight(double weight) { this.weight = weight; }
 
-    public BigDecimal getHeight() { return height; }
-    public void setHeight(BigDecimal height) { this.height = height; }
+    public double getHeight() { return height; }
+    public void setHeight(double height) { this.height = height; }
 
-    public BigDecimal getWingspan() { return wingspan; }
-    public void setWingspan(BigDecimal wingspan) { this.wingspan = wingspan; }
+    public double getWingspan() { return wingspan; }
+    public void setWingspan(double wingspan) { this.wingspan = wingspan; }
 
-    public BigDecimal getVertical() { return vertical; }
-    public void setVertical(BigDecimal vertical) { this.vertical = vertical; }
+    public double getVertical() { return vertical; }
+    public void setVertical(double vertical) { this.vertical = vertical; }
 
     public LocalDate getDateRequested() { return dateRequested; }
     public void setDateRequested(LocalDate dateRequested) { this.dateRequested = dateRequested; }
