+++ backend/CourtVision/src/main/java/cit/edu/capstone/CourtVision/entity/PhysicalUpdateRequest.java
@@ -28,10 +28,10 @@ public class PhysicalUpdateRequest {
     @JsonIgnore
     private List<PhysicalUpdateRequestCoach> requestCoaches;
 
-    private BigDecimal weight;
-    private BigDecimal height;
-    private BigDecimal wingspan;
-    private BigDecimal vertical;
+    private double weight;
+    private double height;
+    private double wingspan;
+    private double vertical;
     private LocalDate dateRequested;
     private int requestStatus; // 0: Pending, 1: Approved, 2: Rejected
 
@@ -51,17 +51,17 @@ public class PhysicalUpdateRequest {
     public List<PhysicalUpdateRequestCoach> getRequestCoaches() { return requestCoaches; }
     public void setRequestCoaches(List<PhysicalUpdateRequestCoach> requestCoaches) { this.requestCoaches = requestCoaches; }
 
-    public BigDecimal getWeight() { return weight; }
-    public void setWeight(BigDecimal weight) { this.weight = weight; }
+    public double getWeight() { return weight; }
+    public void setWeight(double weight) { this.weight = weight; }
 
-    public BigDecimal getHeight() { return height; }
-    public void setHeight(BigDecimal height) { this.height = height; }
+    public double getHeight() { return height; }
+    public void setHeight(double height) { this.height = height; }
 
-    public BigDecimal getWingspan() { return wingspan; }
-    public void setWingspan(BigDecimal wingspan) { this.wingspan = wingspan; }
+    public double getWingspan() { return wingspan; }
+    public void setWingspan(double wingspan) { this.wingspan = wingspan; }
 
-    public BigDecimal getVertical() { return vertical; }
-    public void setVertical(BigDecimal vertical) { this.vertical = vertical; }
+    public double getVertical() { return vertical; }
+    public void setVertical(double vertical) { this.vertical = vertical; }
 
     public LocalDate getDateRequested() { return dateRequested; }
     public void setDateRequested(LocalDate dateRequested) { this.dateRequested = dateRequested; }
